requirements.txt
Django==4.2.19
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
selenium==4.28.1
langchain==0.3.19
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-community==0.3.18
google-genai>=1.0.0
fastmcp>=2.0.0
httpx>=0.27.0
